Creating seperate component for shared styles.
