+++ graylog2-web-interface/src/views/components/searchbar/SearchButtonAndQueryContianer.tsx
@@ -1,0 +1,10 @@
+import styled from 'styled-components';
+
+const SearchButtonAndQuery = styled.div`
+  flex: 1;
+  display: flex;
+  align-items: flex-start;
+  gap: 10px;
+`;
+
+export default SearchButtonAndQuery;
