+++ graylog2-web-interface/src/views/components/searchbar/SearchInputAndValidationContainer.tsx
@@ -1,0 +1,8 @@
+import styled from 'styled-components';
+
+const SearchInputAndValidationContainer = styled.div`
+  display: flex;
+  flex: 1;
+`;
+
+export default SearchInputAndValidationContainer;
